feat: add PlatformSwitch to home page
